+++ app/javascript/src/apps/mydb/elements/details/reactions/variationsTab/ReactionVariationsComponents.js
